Move shared_core to own module and add as feature
